pyproject.toml
[project]
name = "dokiscripts-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "ruff>=0.15.7",
    "tqdm>=4.67.3",
    "zstandard>=0.25.0",
]

[tool.ruff]
target-version = "py312"
line-length = 140
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
# ignoring E501 disables line length check for ruff since we are using a formatter.
ignore = ["E501"]
# Allow fix for all enabled rules (when `--fix` is provided).
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
